Moved homepage container
