+++ src/js/containers/homepage/CovidHighlights.jsx
